requirements.txt
PyMuPdf==1.24.5
streamlit==1.35.0
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.2.3
faiss-cpu==1.8.0
huggingface_hub==0.23.3
sentence-transformers
Pillow==9.5.0
opencv-python==4.7.0.72
pytesseract
numpy
langchain-core
transformers==4.28.1
langchain_community